Fast-Api/requirements.txt
pydantic
fastapi
google.cloud.firestore
uvicorn